NonSmoothSolvers

Documentation for NonSmoothSolvers.

NonSmoothSolvers.GradientSampling

NonSmoothSolvers.OptimizationState

NonSmoothSolvers.OptimizerParams

NonSmoothSolvers.linesearch_nsbfgs

NonSmoothSolvers.optimize!

NonSmoothSolvers.update_iterate!

NonSmoothSolvers.GradientSampling — Type

Gradient sampling algorthm.

NonSmoothSolvers.OptimizationState — Type

OptimizationState

Stores information after one iteration of the optimizer. Generic information is stored explicitly in the struct, custom information may be stored in the field additionalinfo::NamedTuple.

NonSmoothSolvers.OptimizerParams — Type

OptimizerParams

Generic parameters for optim algs

NonSmoothSolvers.linesearch_nsbfgs — Method

linesearch_nsbfgs

Nonsmooth linesearch from Nonsmooth optimization via quasi-Newton methods, Lewis & Overton, 2013.

NonSmoothSolvers.optimize! — Method

optimize!(pb, optimizer::NonSmoothSolvers.Optimizer, initial_x; state, optimstate_extensions, optparams) -> Tuple{Any, Any}

Call the optimizer on problem pb, with initial point initial_x. Returns a tuple containing the final iterate vector and a trace.

Features:

timing of the update_iterate method only;

saves basic information of each iteration in a vector of OptimizationState, the so-called trace;

the information saved at each iterate may be enriched by the user by providing a name and callback function via the optimstate_extension argument.

Example

getx(o, os) = os.x
optimstate_extensions = OrderedDict{Symbol, Function}(:x => getx)

optimize!(pb, o, xclose; optparams, optimstate_extensions)

NonSmoothSolvers.update_iterate! — Method

update_iterate!(state, gs::GradientSampling, pb)

NOTE: each iteration is costly. This can be explored with NonSmoothProblems.to. On the maxquadBGLS problem ──────────────────────────────────────────────────────────────────────────────────────────────────── Time Allocations ─────────────────────── ──────────────────────── Tot / % measured: 103ms / 93.9% 9.50MiB / 96.1%

Section ncalls time %tot avg alloc %tot avg ──────────────────────────────────────────────────────────────────────────────────────────────────── updateiterate! 100 95.6ms 98.7% 956μs 8.92MiB 97.8% 91.4KiB GS 2. minimum norm (sub)gradient 100 83.3ms 86.0% 833μs 3.56MiB 39.0% 36.4KiB GS 1. sampling points, eval gradients 100 7.39ms 7.6% 73.9μs 3.85MiB 42.2% 39.4KiB GS 4. Update parameters 100 4.24ms 4.4% 42.4μs 1.40MiB 15.3% 14.3KiB GS 5. diff check 100 288μs 0.3% 2.88μs 117KiB 1.3% 1.17KiB GS 3. Termination 100 41.1μs 0.0% 411ns 0.00B 0.0% 0.00B buildoptimstate 100 1.19ms 1.2% 11.9μs 207KiB 2.2% 2.07KiB CV check 100 47.0μs 0.0% 470ns 0.00B 0.0% 0.00B ────────────────────────────────────────────────────────────────────────────────────────────────────